feat(blog): add new blog post about a recently published paper on multimodal large language model (MLLM)
The blog post discusses a paper called "Language Is Not All You Need: Aligning Perception with Language Models" that introduces an architecture that is able to perceive two modalities, or in other words, multimodal large language model (MLLM) called Kosmos-1, which is a necessary step towards Artificial General Intelligence (AGI).

The post explains the concept of multi-modality and AGI, and how the authors applied the idea of using Language Models as a general-purpose-interface to other various foundation models. The post also discusses the architecture of Kosmos-1, which uses visual transformers (ViT) to transform an image to a form that can be understood by the general-purpose-interface (Transformer Decoder). The post also evaluates the model using a lot of different techniques, including the evaluation of its fluid reasoning ability of humans using Raven’s Progressive Matrices.

The